+++ packages/kbn-i18n/angular/package.json
@@ -1,0 +1,4 @@
+{
+  "browser": "../target/web/angular",
+  "main": "../target/node/angular"
+}
