+++ packages/kbn-i18n/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@kbn/i18n",
-  "browser": "./target/web/browser.js",
-  "main": "./target/node/index.js",
+  "browser": "./target/web",
+  "main": "./target/node",
   "module": "./src/index.js",
   "version": "1.0.0",
   "license": "Apache-2.0",
