+++ packages/kbn-i18n/src/react/provider.js
@@ -21,7 +21,7 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { IntlProvider } from 'react-intl';
 
-import * as i18n from '../core/i18n';
+import * as i18n from '../core';
 
 /**
  * The library uses the provider pattern to scope an i18n context to a tree
